Move definition of -D__ARM_FEATURE_LPAE=1 cflag to top-level.
LPAE indicates better instructions can be used when atomicity guarantees are
needed. However, LPAE's presence isn't advertised by clang/GCC. We fake an
ARM feature to advertise its presence on architectures where it is.
Also, add a TODO documenting that cortex-a15 is not the correct CPU variant
for krait.

Change-Id: I580ceea16b2f176543bf3d0ba8d8d998cc1c6bdb
Signed-off-by: Maxime Poulain <maxime.poulain@student.uclouvain.be>

Conflicts:
	core/combo/arch/arm/armv7-a-neon.mk